Move common types to src/@types directory
