+++ src/redux/chromeReducers.ts
@@ -1,7 +1,7 @@
 import { QuickStart } from '@patternfly/quickstarts';
 import { ChromeUser } from '@redhat-cloud-services/types';
 import { REQUESTS_COUNT, REQUESTS_DATA } from '../utils/consts';
-import { NavItem } from '../js/types';
+import { NavItem } from '../@types/types';
 import { generateRoutesList, highlightItems, isBeta, isFedRamp, levelArray } from '../utils/common';
 import { ThreeScaleError } from '../utils/responseInterceptors';
 import { AccessRequest, ChromeModule, ChromeState, Navigation } from './store';
