+++ src/utils/consts.ts
@@ -5,7 +5,7 @@ import cookie from 'js-cookie';
 import { getFeatureFlagsError, unleashClient } from '../js/App/FeatureFlags/FeatureFlagsProvider';
 import { isBeta, isProd } from './common';
 import { AxiosRequestConfig } from 'axios';
-import { ChromeAuthOptions } from '../js/types';
+import { ChromeAuthOptions } from '../@types/types';
 
 export const noAuthParam = 'noauth';
 export const offlineToken = '2402500adeacc30eb5c5a8a5e2e0ec1f';
